+++ R/estimate_vpd_from_dew.R
@@ -39,7 +39,7 @@ sEddyProc$methods(sFillVPDFromDew =
 #'
 #' Since sometimes Esat_daymin is lower than Esat(Tair_min)
 #' the estimated VPDfromDew is underestimated. This function applies a
-#' linear model of the exisitng VPD and estimated VPD to correct for this bias:
+#' linear model of the existing VPD and estimated VPD to correct for this bias:
 #' VPD ~ 0 + VPDfromDew * Tair_f * hourOfDay *  TminOftheDay * TRangeDay
 #'
 #' @param df data.frame with columns DateTime, VPD, Tair, and Tair_f
@@ -72,12 +72,12 @@ estimate_vpd_from_dew <- function(df, pNonMissing = 0.1){
       TmaxOftheDay = suppressWarnings(max(.data$Tair_f, na.rm = TRUE)),
       TRangeDay = .data$TmaxOftheDay - .data$TminOftheDay,
       EminDew = Esat.slope(.data$TminOftheDay)$Esat*10,
-      VPDfromDew = Esat.slope(.data$Tair_f)$Esat*10 - .data$EminDew 
+      VPDfromDew = Esat.slope(.data$Tair_f)$Esat*10 - .data$EminDew
     ) %>%
     ungroup()
   # VPDfromDew is computed using Tair_f, but for the correction only use
   # those cases where original Tair is finite to avoid confounding issues
-  lm1 <- lm(VPD ~ 0 + VPDfromDew * Tair_f * hourOfDay *  TminOftheDay * TRangeDay, 
+  lm1 <- lm(VPD ~ 0 + VPDfromDew * Tair_f * hourOfDay *  TminOftheDay * TRangeDay,
             filter(df_f, is.finite(.data$Tair)))
   # create column in original data.frame (do not return the intermediate vars)
   # the order in df and df_f should not have changed with grouping/ungrouping